+++ src/Careerbuilder.php
@@ -5,32 +5,214 @@ use JobBrander\Jobs\Client\Job;
 class Careerbuilder extends AbstractProvider
 {
     /**
-     * Developer Key
+     * Default return attributes
      *
-     * @var string
+     * @var array
      */
-    protected $developerKey;
+    protected $defaultAttributes = [
+        'Company',
+        'CompanyDetailsURL',
+        'DescriptionTeaser',
+        'DID',
+        'OnetCode',
+        'ONetFriendlyTitle',
+        'EmploymentType',
+        'EducationRequired',
+        'ExperienceRequired',
+        'JobDetailsURL',
+        'Location',
+        'City',
+        'State',
+        'PostedDate',
+        'Pay',
+        'JobTitle',
+        'CompanyImageURL',
+        'Skills',
+    ];
 
     /**
-     * Host Site
+     * Map of setter methods to query parameters
      *
-     * @var string
+     * @var array
      */
-    protected $hostSite;
+    protected $queryMap = [
+        'setDeveloperKey' => 'DeveloperKey',
+        'setAdvancedGroupingMode' => 'AdvancedGroupingMode',
+        'setApplyRequirements' => 'ApplyRequirements',
+        'setBooleanOperator' => 'BooleanOperator',
+        'setCategory' => 'Category',
+        'setCity' => 'FacetCity',
+        'setCoBrand' => 'CoBrand',
+        'setCompanyDID' => 'CompanyDID',
+        'setCompanyDIDCSV' => 'CompanyDIDCSV',
+        'setCompanyName' => 'CompanyName',
+        'setCompanyNameBoostParams' => 'CompanyNameBoostParams',
+        'setCount' => 'PerPage',
+        'setCountryCode' => 'CountryCode',
+        'setEducationCode' => 'EducationCode',
+        'setEmpType' => 'EmpType',
+        'setEnableCompanyCollapse' => 'EnableCompanyCollapse',
+        'setEnableCompanyJobTitleCollapse' => 'EnableCompanyJobTitleCollapse',
+        'setExcludeApplyRequirements' => 'ExcludeApplyRequirements',
+        'setExcludeCompanyNames' => 'ExcludeCompanyNames',
+        'setExcludeJobTitles' => 'ExcludeJobTitles',
+        'setExcludeKeywords' => 'ExcludeKeywords',
+        'setExcludeNational' => 'ExcludeNational',
+        'setExcludeNonTraditionalJobs' => 'ExcludeNonTraditionalJobs',
+        'setFacetCategory' => 'FacetCategory',
+        'setFacetCompany' => 'FacetCompany',
+        'setFacetCity' => 'FacetCity',
+        'setFacetState' => 'FacetState',
+        'setFacetCityState' => 'FacetCityState',
+        'setFacetPay' => 'FacetPay',
+        'setFacetRelatedJobTitle' => 'FacetRelatedJobTitle',
+        'setFacetCountry' => 'FacetCountry',
+        'setFacetEmploymentType' => 'FacetEmploymentType',
+        'setHostSite' => 'HostSite',
+        'setIncludeCompanyChildren' => 'IncludeCompanyChildren',
+        'setIndustryCodes' => 'IndustryCodes',
+        'setJobTitle' => 'JobTitle',
+        'setKeywords' => 'Keywords',
+        'setKeyword' => 'Keywords',
+        'setLocation' => 'Location',
+        'setNormalizedCompanyDID' => 'NormalizedCompanyDID',
+        'setNormalizedCompanyDIDBoostParams' => 'NormalizedCompanyDIDBoostParams',
+        'setNormalizedCompanyName' => 'NormalizedCompanyName',
+        'setONetCode' => 'ONetCode',
+        'setOrderBy' => 'OrderBy',
+        'setOrderDirection' => 'OrderDirection',
+        'setPage' => 'PageNumber',
+        'setPageNumber' => 'PageNumber',
+        'setPartnerID' => 'PartnerID',
+        'setPayHigh' => 'PayHigh',
+        'setPayInfoOnly' => 'PayInfoOnly',
+        'setPayLow' => 'PayLow',
+        'setPerPage' => 'PerPage',
+        'setPostedWithin' => 'PostedWithin',
+        'setRadius' => 'Radius',
+        'setRecordsPerGroup' => 'RecordsPerGroup',
+        'setRelocateJobs' => 'RelocateJobs',
+        'setSOCCode' => 'SOCCode',
+        'setSchoolSiteID' => 'SchoolSiteID',
+        'setSearchAllCountries' => 'SearchAllCountries',
+        'setSearchView' => 'SearchView',
+        'setShowCategories' => 'ShowCategories',
+        'setShowApplyRequirements' => 'ShowApplyRequirements',
+        'setSingleONetSearch' => 'SingleONetSearch',
+        'setSiteEntity' => 'SiteEntity',
+        'setSiteID' => 'SiteID',
+        'setSkills' => 'Skills',
+        'setSpecificEducation' => 'SpecificEducation',
+        'setSpokenLanguage' => 'SpokenLanguage',
+        'setState' => 'FacetState',
+        'setTags' => 'Tags',
+        'setTalentNetworkDID' => 'TalentNetworkDID',
+        'setUrlCompressionService' => 'UrlCompressionService',
+        'setUseFacets' => 'UseFacets',
+    ];
 
     /**
-     * Use Facets (for city/state)
+     * Current api query parameters
      *
-     * @var string
+     * @var array
      */
-    protected $useFacets;
+    protected $queryParams = [
+        'DeveloperKey' => null,
+        'AdvancedGroupingMode' => null,
+        'ApplyRequirements' => null,
+        'BooleanOperator' => null,
+        'Category' => null,
+        'CoBrand' => null,
+        'CompanyDID' => null,
+        'CompanyDIDCSV' => null,
+        'CompanyName' => null,
+        'CompanyNameBoostParams' => null,
+        'CountryCode' => null,
+        'EducationCode' => null,
+        'EmpType' => null,
+        'EnableCompanyCollapse' => 'true',
+        'EnableCompanyJobTitleCollapse' => null,
+        'ExcludeApplyRequirements' => null,
+        'ExcludeCompanyNames' => null,
+        'ExcludeJobTitles' => null,
+        'ExcludeKeywords' => null,
+        'ExcludeNational' => null,
+        'ExcludeNonTraditionalJobs' => null,
+        'FacetCategory' => null,
+        'FacetCompany' => null,
+        'FacetCity' => null,
+        'FacetState' => null,
+        'FacetCityState' => null,
+        'FacetPay' => null,
+        'FacetRelatedJobTitle' => null,
+        'FacetCountry' => null,
+        'FacetEmploymentType' => null,
+        'HostSite' => 'US',
+        'IncludeCompanyChildren' => null,
+        'IndustryCodes' => null,
+        'JobTitle' => null,
+        'Keywords' => null,
+        'Location' => null,
+        'NormalizedCompanyDID' => null,
+        'NormalizedCompanyDIDBoostParams' => null,
+        'NormalizedCompanyName' => null,
+        'ONetCode' => null,
+        'OrderBy' => null,
+        'OrderDirection' => null,
+        'PageNumber' => null,
+        'PartnerID' => null,
+        'PayHigh' => null,
+        'PayInfoOnly' => null,
+        'PayLow' => null,
+        'PerPage' => null,
+        'PostedWithin' => null,
+        'Radius' => null,
+        'RecordsPerGroup' => null,
+        'RelocateJobs' => null,
+        'SOCCode' => null,
+        'SchoolSiteID' => null,
+        'SearchAllCountries' => null,
+        'SearchView' => null,
+        'ShowCategories' => null,
+        'ShowApplyRequirements' => null,
+        'SingleONetSearch' => null,
+        'SiteEntity' => null,
+        'SiteID' => null,
+        'Skills' => null,
+        'SpecificEducation' => null,
+        'SpokenLanguage' => null,
+        'Tags' => null,
+        'TalentNetworkDID' => null,
+        'UrlCompressionService' => null,
+        'UseFacets' => 'true',
+    ];
 
     /**
-     * Enable Company Collapse (must be true to set count)
+     * Create new Careerbuilder jobs client.
      *
-     * @var string
+     * @param array $parameters
      */
-    protected $companyCollapse;
+    public function __construct($parameters = [])
+    {
+        parent::__construct($parameters);
+        array_walk($parameters, [$this, 'updateQuery']);
+    }
+
+    /**
+     * Magic method to handle get and set methods for properties
+     *
+     * @param  string $method
+     * @param  array  $parameters
+     *
+     * @return mixed
+     */
+    public function __call($method, $parameters)
+    {
+        if (isset($this->queryMap[$method], $parameters[0])) {
+            $this->updateQuery($parameters[0], $this->queryMap[$method]);
+        }
+        return parent::__call($method, $parameters);
+    }
 
     /**
      * Returns the standardized job object
@@ -41,31 +223,9 @@ class Careerbuilder extends AbstractProvider
      */
     public function createJobObject($payload = [])
     {
-        $defaults = [
-            'Company',
-            'CompanyDetailsURL',
-            'DescriptionTeaser',
-            'DID',
-            'OnetCode',
-            'ONetFriendlyTitle',
-            'EmploymentType',
-            'EducationRequired',
-            'ExperienceRequired',
-            'JobDetailsURL',
-            'Location',
-            'City',
-            'State',
-            'PostedDate',
-            'Pay',
-            'JobTitle',
-            'CompanyImageURL',
-            'Skills',
-        ];
+        $payload = static::parseAttributeDefaults($payload, $this->defaultAttributes);
 
-        $payload = static::parseAttributeDefaults($payload, $defaults);
-
-        $job = new Job(
-            [
+        $job = new Job([
             'description' => $payload['DescriptionTeaser'],
             'employmentType' => $payload['EmploymentType'],
             'title' => $payload['JobTitle'],
@@ -74,8 +234,7 @@ class Careerbuilder extends AbstractProvider
             'educationRequirements' => $payload['EducationRequired'],
             'experienceRequirements' => $payload['ExperienceRequired'],
             'sourceId' => $payload['DID'],
-            ]
-        );
+        ]);
 
         $pay = $this->parseSalariesFromString($payload['Pay']);
 
@@ -93,40 +252,10 @@ class Careerbuilder extends AbstractProvider
             ->setMaximumSalary($pay['max']);
 
         if (isset($payload['Skills']['Skill'])) {
-            $job->setSkills($this->getSkillSet($payload['Skills']['Skill']));
+            $job->setSkills($this->parseSkillSet($payload['Skills']['Skill']));
         }
 
         return $job;
-    }
-
-    /**
-     * Get host site
-     *
-     * @return string
-     */
-    public function getHostSite()
-    {
-        return 'US';
-    }
-
-    /**
-     * Get Use Facets
-     *
-     * @return string
-     */
-    public function getUseFacets()
-    {
-        return 'true';
-    }
-
-    /**
-     * Get Enable Company Collapse
-     *
-     * @return string
-     */
-    public function getCompanyCollapse()
-    {
-        return 'true';
     }
 
     /**
@@ -156,30 +285,28 @@ class Careerbuilder extends AbstractProvider
      */
     public function getQueryString()
     {
-        $queryParams = [
-            'DeveloperKey' => 'getDeveloperKey',
-            'Keywords' => 'getKeyword',
-            'FacetState' => 'getState',
-            'FacetCity' => 'getCity',
-            'PageNumber' => 'getPage',
-            'PerPage' => 'getCount',
-            'UseFacets' => 'getUseFacets',
-            'EnableCompanyCollapse' => 'getCompanyCollapse',
-        ];
+        return http_build_query($this->queryParams);
+    }
 
-        $queryString = [];
+    /**
+     * Get url
+     *
+     * @return string
+     */
+    public function getUrl()
+    {
+        $queryString = $this->getQueryString();
+        return 'http://api.careerbuilder.com/v2/jobsearch/?'.$queryString;
+    }
 
-        array_walk(
-            $queryParams,
-            function ($value, $key) use (&$queryString) {
-                $computedValue = $this->$value();
-                if (!is_null($computedValue)) {
-                    $queryString[$key] = $computedValue;
-                }
-            }
-        );
-
-        return http_build_query($queryString);
+    /**
+     * Get http verb
+     *
+     * @return string
+     */
+    public function getVerb()
+    {
+        return 'GET';
     }
 
     /**
@@ -210,7 +337,12 @@ class Careerbuilder extends AbstractProvider
         return $salary;
     }
 
-    public function parseAnnualRange($salary = [], $input = null)
+    /**
+     * Parse annual salary range from CB API
+     *
+     * @return array
+     */
+    protected function parseAnnualRange($salary = [], $input = null)
     {
         preg_replace_callback("/(.\d+k)\s.\s(.\d+k)/", function ($matches) use (&$salary) {
             $salary['min'] = str_replace('k', '000', $matches[1]);
@@ -220,7 +352,12 @@ class Careerbuilder extends AbstractProvider
         return $salary;
     }
 
-    public function parseAnnualFixed($salary = [], $input = null)
+    /**
+     * Parse fixed annual salary from CB API
+     *
+     * @return array
+     */
+    protected function parseAnnualFixed($salary = [], $input = null)
     {
         preg_replace_callback("/(.\d+k)/", function ($matches) use (&$salary) {
             $salary['min'] = str_replace('k', '000', $matches[1]);
@@ -229,7 +366,12 @@ class Careerbuilder extends AbstractProvider
         return $salary;
     }
 
-    public function parseHourlyRange($salary = [], $input = null)
+    /**
+     * Parse hourly payrate range from CB API
+     *
+     * @return array
+     */
+    protected function parseHourlyRange($salary = [], $input = null)
     {
         preg_replace_callback("/(.\d+.\d+)\s.\s(.\d+.\d+)/", function ($matches) use (&$salary) {
             $salary['min'] = $matches[1];
@@ -239,7 +381,12 @@ class Careerbuilder extends AbstractProvider
         return $salary;
     }
 
-    public function parseHourlyFixed($salary = [], $input = null)
+    /**
+     * Parse fixed hourly payrate from CB API
+     *
+     * @return array
+     */
+    protected function parseHourlyFixed($salary = [], $input = null)
     {
         preg_replace_callback("/(.\d+.\d+)/", function ($matches) use (&$salary) {
             $salary['min'] = $matches[1];
@@ -249,28 +396,11 @@ class Careerbuilder extends AbstractProvider
     }
 
     /**
-     * Get url
+     * Parse skills array into string
      *
-     * @return string
+     * @return array
      */
-    public function getUrl()
-    {
-        $queryString = $this->getQueryString();
-
-        return 'http://api.careerbuilder.com/v2/jobsearch/?'.$queryString;
-    }
-
-    /**
-     * Get http verb
-     *
-     * @return string
-     */
-    public function getVerb()
-    {
-        return 'GET';
-    }
-
-    public function getSkillSet($skills)
+    protected function parseSkillSet($skills)
     {
         if (is_array($skills)) {
             return implode(', ', $skills);
@@ -278,5 +408,21 @@ class Careerbuilder extends AbstractProvider
             return $skills;
         }
         return null;
+    }
+
+    /**
+     * Attempts to update current query parameters.
+     *
+     * @param  string  $value
+     * @param  string  $key
+     *
+     * @return Careerbuilder
+     */
+    protected function updateQuery($value, $key)
+    {
+        if (array_key_exists($key, $this->queryParams)) {
+            $this->queryParams[$key] = $value;
+        }
+        return $this;
     }
 }