+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 # Changelog
 All Notable changes to `jobs-careerbuilder` will be documented in this file
 
+## 1.0.0 - 2015-09-27
+
+### Added
+- Support for all setter methods outlined in the [Careerbuilder API](http://api.careerbuilder.com/Search/jobsearch/jobsearchinfo.aspx)
+- Readme documentation for all supported methods
+- Tests to support all new methods
+
+### Deprecated
+- Public methods for parsing range/fixed salaries made protected
+
+### Fixed
+- Sorting methods alphabetically
+- Travis-ci support for PHP 7.0 and HHVM
+
+### Removed
+- Public "getSkillSet" replaced with protected "parseSkillSet" method
+
+### Security
+- Nothing
+
 ## 0.2.0 - 2015-09-27
 
 ### Added